// types.d.ts
export type Product = {
  id: number;
  title: string;
  price: number;
  description: string;
  images: string[];
  creationAt: string;
  updatedAt: string;
  category: {
    id: number;
    name: string;
    image: string;
    creationAt: string;
    updatedAt: string;
  };
};

export type RequestProducts = {
  products: Product[];
};

export type Imagestype = string[];

export type Categories = {
  id: number;
  name: string;
  image: string;
  creationAt: string;
  updatedAt: string;
};